pyproject.toml
[project]
name = "stamp-cmp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "paddleocr>=3.3.3",
    "paddlepaddle>=3.3.0",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "timm>=1.0.24",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "trdg>=0.1.1",
    "uvicorn>=0.40.0",
]
